Output is now normalized
Fixed syntax problems with cells, matrices, and arrays. The files are now normailzed on output. Setting all weights to 1 in the vector of weights makes the std() of each wav 1.

The syntax for inputting the weights is
sig = wav2sig(fnames,fs,tInt,weights)

Next commit will be adding option of this syntax
sig = wav2sig(fnames,fs,weights)
and this
sig = wav2sig(fnames,weights)
for the scaled RMS processing.