Struct Input, Fixed all Dependent Files
I changed all input arguments to one structure. Having so many optional inputs made it very cumbersome to sort out what was what, and forced some unnessary code. fnames is still the only required argument, and varargin is still used for the rest of the arguments.

A structure containing any combination of optional fields 'fs','tInt', and 'weights' is optional, but is expected to be the only item in varargin. Using a structure shrunk down the code nearly 100 lines.

All files that I could find that called wav2sig were changed such that they wav2sig with the structure where needed. I tested all dependent files.

Fixed a bug or two in dependent files as well.